updating namespace and pushing to release
