+++ pyvfx/boilerplate/mayapalette.py
