+++ pyvfx/boilerplate/resources/main_window.ui
