+++ pyvfx/boilerplate/resources/module.ui
