+++ pyvfx/boilerplate/resources/qpalette_maya2015.json
