+++ pyvfx/boilerplate/resources/qpalette_maya2016.json
