+++ pyvfx/boilerplate/utils/get_set_palette_data_qt4.py
